Cleaning server.py, moving initializing method to separate initialize.py
