+++ APIStockServer/server.py
@@ -2,50 +2,35 @@ import time
 import os
 
 from APIStockServer.Alerts import MetalEtfAlerts, AlertsScheduler
-from APIStockServer.Alerts.AlertSender import EmailSender
-from APIStockServer.Alerts.MetalEtfCalculation import SprottEtfCalculation
 from APIStockServer.Alerts.MetalEtfCalculation import PhysEtfCalculation, PslvEtfCalculation, SpppEtfCalculation
-from APIStockServer.DataAcquisition import Finnhub, GoldApi, MetalEtfScraping
+from APIStockServer.DataAcquisition import MetalEtfScraping
 from APIStockServer.modules.config_file import ConfigFile
-
-
-def initialize_api_objects(config: ConfigFile):
-    finnhub_keys = config.read_file_defined_by_key("API Keys", "finnhub")
-    finnhub_apis = list(Finnhub(key) for key in finnhub_keys)
-    goldapi_keys = config.read_file_defined_by_key("API Keys", "goldapi")
-    goldapi_apis = list(GoldApi(key) for key in goldapi_keys)
-    return finnhub_apis, goldapi_apis
-
-
-def initialize_email_sender(config: ConfigFile):
-    config_email = config['Email Sender']
-    email_sender = EmailSender(config_email['smtp_server'], config_email['sender_email'], config_email['password'])
-    return email_sender
+from APIStockServer.initialize import initialize_api_objects, initialize_email_sender, initialize_sprott_etf_calculators
 
 
 if __name__ == "__main__":
+    # Initialize object handling config file
     config_obj = ConfigFile(os.path.join(os.path.dirname(__file__), "config.ini"))
 
+    # Initialize API objects (Finnhub, GoldAPI)
     finnhub_apis, goldapi_apis = initialize_api_objects(config_obj)
 
+    # Initialize Email Sender object
     email_sender = initialize_email_sender(config_obj)
-
+    # Read email addresses for email_sender from config file
     receivers_emails = config_obj.read_file_defined_by_key("Email Sender", "receiver_emails_path")
 
+    # Initialize Sprott Etfs discount calculators
     list_of_sprott_etfs = [PhysEtfCalculation, PslvEtfCalculation, SpppEtfCalculation]
-    calculators = list()
-    for C in list_of_sprott_etfs:
-        C: SprottEtfCalculation
-        calculators.append(C(spot_data=goldapi_apis[0],
-                             etf_data=finnhub_apis[0],
-                             etf_info=MetalEtfScraping()
-                             )
-                           )
+    sprott_etfs_calculators = initialize_sprott_etf_calculators(list_of_sprott_etfs,
+                                                                goldapi_apis[0],
+                                                                finnhub_apis[0],
+                                                                MetalEtfScraping())
 
-    metal_etf_alerts = MetalEtfAlerts(email_sender, calculators, receivers_emails)
+    metal_etf_alerts = MetalEtfAlerts(email_sender, sprott_etfs_calculators, receivers_emails)
     AlertsScheduler(metal_etf_alerts.check_metal_alerts, print_log=True,
                     start_hour=10, start_minute=30,
-                    stop_hour=21, stop_minute=50,
+                    stop_hour=1, stop_minute=50,
                     )
     while True:
         AlertsScheduler.run_waiting_alert()
